+++ docs/src/examples/QasGridGenerator/WithFieldsProps.vue
@@ -1,6 +1,6 @@
 <template>
   <!-- Usando qas-single-view apenas para recuperar os dados -->
-  <qas-single-view v-model:fields="fields" v-model:result="result" :custom-id="customId" :entity="entity">
+  <qas-single-view v-model:fields="viewState.fields" v-model:result="viewState.result" :custom-id :entity>
     <template #default>
       <qas-grid-generator v-bind="gridGeneratorProps" />
     </template>
@@ -8,12 +8,12 @@
 </template>
 
 <script setup>
-import { ref, computed } from 'vue'
+import { useView } from '@bildvitta/composables'
+import { computed } from 'vue'
 
 defineOptions({ name: 'WithFieldsProps' })
 
-const fields = ref({})
-const result = ref({})
+const { viewState } = useView({ mode: 'single' })
 
 const entity = 'users'
 
@@ -22,8 +22,8 @@ const customId = '3102fad5-f14c-45d4-98e9-46ef0aa9580e'
 
 const gridGeneratorProps = computed(() => {
   return {
-    fields: fields.value,
-    result: result.value,
+    fields: viewState.value.fields,
+    result: viewState.value.result,
     useBox: true,
     useInline: true,