+++ docs/src/examples/QasGridItem/WithCustomTypography.vue
@@ -1,11 +1,28 @@
 <template>
   <div class="container q-py-lg">
-    <qas-grid-item :content="{ typography: 'h4' }" label="Nome" use-inline value="John doe" />
+    <qas-grid-item v-bind="firstGridItemProps" />
 
-    <qas-grid-item class="q-mt-lg" :content="{ typography: 'h4' }" label="Nome" value="John doe" />
+    <qas-grid-item class="q-mt-lg" v-bind="secondGridItemProps" />
   </div>
 </template>
 
 <script setup>
 defineOptions({ name: 'WithCustomTypography' })
+
+const firstGridItemProps = {
+  content: {
+    typography: 'h4'
+  },
+  label: 'Nome',
+  useInline: true,
+  value: 'John doe'
+}
+
+const secondGridItemProps = {
+  content: {
+    typography: 'h4'
+  },
+  label: 'Nome',
+  value: 'John doe'
+}
 </script>